created allocator

+++ readfile.h
@@ -1,6 +1,6 @@
 #ifndef READFILE_H
 #define READFILE_H
 
-int readfile(char **buff, char *filename);
+long readfile(char **buff, char *filename);
 
 #endif
